# docs/schema/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "ConstraintPoidsUpdate" : {
        "type" : "object",
        "properties" : {
          "poids" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ConstraintToggleUpdate" : {
        "required" : [ "actif" ],
        "type" : "object",
        "properties" : {
          "actif" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  },
  "paths" : {
...
    "/api/constraints/{name}" : {
      "put" : {
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConstraintToggleUpdate"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConstraintToggleUpdate"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Set Actif",
        "tags" : [ "Constraint Resource" ]
      }
    },
    "/api/constraints/{name}/poids" : {
      "put" : {
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConstraintPoidsUpdate"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConstraintPoidsUpdate"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Set Poids",
        "tags" : [ "Constraint Resource" ]
      }
    },
...
  },
  "info" : {
    "title" : "planning-equipes API",
    "version" : "999-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}